refactor: fetcher to use generic type
