+++ src/app/app.server.module.ts
@@ -4,7 +4,7 @@ import { ServerModule } from '@angular/platform-server'
 import { AppModule } from './app.module'
 import { AppComponent } from './app.component'
 import { LocalJsonFetcherService } from './common/json-fetcher/local-json-fetcher.service'
-import { JsonFetcher } from './common/json-fetcher/json-fetcher-injection-token'
+import { JsonFetcher } from './common/json-fetcher/json-fetcher'
 
 @NgModule({
   imports: [AppModule, ServerModule],
