+++ src/app/common/json-fetcher/http-json-fetcher.service.ts
@@ -1,9 +1,8 @@
-import { JSON_DATA_DIR, JsonFetcher } from './json-fetcher-injection-token'
+import { JSON_DATA_DIR, JsonFetcher } from './json-fetcher'
 import { HttpClient } from '@angular/common/http'
 import { firstValueFrom } from 'rxjs'
 import { Inject, Injectable } from '@angular/core'
 import { APP_BASE_HREF } from '@angular/common'
-import { Json } from './json-types'
 
 @Injectable({ providedIn: 'root' })
 export class HttpJsonFetcherService implements JsonFetcher {
@@ -13,17 +12,13 @@ export class HttpJsonFetcherService implements JsonFetcher {
     @Inject(JSON_DATA_DIR) private jsonDataDir: string,
   ) {}
 
-  async fetch(...pathSegments: string[]): Promise<Json | undefined> {
-    const response = firstValueFrom(
-      this.httpClient.get(
+  async fetch<T>(...pathSegments: string[]): Promise<T> {
+    return firstValueFrom(
+      this.httpClient.get<T>(
         [this.baseHref, this.jsonDataDir, ...pathSegments]
           .join('/')
           .replace(/^\//, ''),
       ),
     )
-    if (!response) {
-      return
-    }
-    return response as unknown as Json
   }
 }
