+++ src/app/common/json-fetcher/json-fetcher.ts
@@ -1,5 +1,4 @@
 import { InjectionToken } from '@angular/core'
-import { Json } from './json-types'
 import { CONTENTS_DIR } from '../directories'
 
 export const JSON_DATA_DIR = new InjectionToken<string>('JSON data dir', {
@@ -7,5 +6,5 @@ export const JSON_DATA_DIR = new InjectionToken<string>('JSON data dir', {
 })
 
 export abstract class JsonFetcher {
-  abstract fetch(...pathSegments: string[]): Promise<Json | undefined>
+  abstract fetch<T>(...pathSegments: string[]): Promise<T>
 }
