+++ src/app/project-page/lookbooks/project-lookbooks.service.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing'
 
 import { ProjectLookbooksService } from './project-lookbooks.service'
 import { MockProviders } from 'ng-mocks'
-import { JsonFetcher } from '../../common/json-fetcher/json-fetcher-injection-token'
+import { JsonFetcher } from '../../common/json-fetcher/json-fetcher'
 
 describe('ProjectLookbooksService', () => {
   let service: ProjectLookbooksService
